TG-141: modify shared table

+++ src/Components/Shared/Table/index.js
@@ -16,29 +16,31 @@ const Table = ({ data, headers, onDelete, onRowClick }) => {
         </thead>
         <tbody className={tableStyles.tbody}>
           {data.map((row) => {
-            return (
-              <>
-                <tr className={tableStyles.tr} key={row._id} onClick={() => onRowClick(row._id)}>
-                  {headers.map((header, index) => {
-                    return (
-                      <>
-                        <td key={index}>{row[header]}</td>
-                      </>
-                    );
-                  })}
-                  <td>
-                    <img
-                      className={tableStyles.logo}
-                      src={Logo}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        onDelete(row._id, true);
-                      }}
-                    />
-                  </td>
-                </tr>
-              </>
-            );
+            if (!row.isDeleted) {
+              return (
+                <>
+                  <tr className={tableStyles.tr} key={row._id} onClick={() => onRowClick(row._id)}>
+                    {headers.map((header, index) => {
+                      return (
+                        <>
+                          <td key={index}>{row[header]}</td>
+                        </>
+                      );
+                    })}
+                    <td>
+                      <img
+                        className={tableStyles.logo}
+                        src={Logo}
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          onDelete(row._id, true);
+                        }}
+                      />
+                    </td>
+                  </tr>
+                </>
+              );
+            }
           })}
         </tbody>
       </table>
